refactor : Banker의 DI, DIP를 위해 별도의 Bankerable 프로토콜을 구현
- Banker을 DI, DIP를 위해 별도의 Bankerable 프로토콜을 생성
- Bankerable프로토콜의 Extension 기본 구현 사용
- Bank 타입 내부 기존의 Banker 타입 사용을 Bankerable로 변경
- Banker 타입 Bankerable 프로토콜 채택
