+++ BankManagerConsoleApp/BankManagerConsoleApp/Models/Bank/Banker.swift
@@ -5,21 +5,4 @@
 //  Created by Kyo, Wonbi on 2022/11/02.
 //
 
-import Foundation
-
-struct Banker {
-    func startWork(client: Client) {
-        print("\(client.waitingTicket)번 고객 \(client.purpose.name)업무 시작")
-        selectWork(purpose: client.purpose)
-        print("\(client.waitingTicket)번 고객 \(client.purpose.name)업무 종료")
-    }
-    
-    private func selectWork(purpose: Client.Purpose) {
-        switch purpose {
-        case .deposit:
-            Thread.sleep(forTimeInterval: 0.7)
-        case .loan:
-            Thread.sleep(forTimeInterval: 1.1)
-        }
-    }
-}
+struct Banker: Bankerable { }
