+++ BankManagerConsoleApp/BankManagerConsoleApp/Protocol/Bankerable.swift
@@ -1,0 +1,8 @@
+//
+//  Bankerable.swift
+//  BankManagerConsoleApp
+//
+//  Created by Kyo, Wonbi on 2022/11/07.
+//
+
+protocol Bankerable { }
